add mac main smooth some parameters
git-svn-id: http://localhost/svn/repository@13 f9e7563d-c07c-ba4f-b3db-62cb7e23badb

+++ OutofRangePeel/OutofrangePeel/Events.h
@@ -6,19 +6,19 @@ class MouseEvent
 private:
 	int button_;
 	int state_;
-	int x_;
-	int y_;
-	int dir_;//wheel
+	float x_;
+	float y_;
+	float dir_;//wheel
 
 public:
 	MouseEvent(){}
-	MouseEvent(int button, int state, int x, int y, int dir):button_(button),state_(state),x_(x),y_(y),dir_(dir){}
+	MouseEvent(int button, int state, float x, float y, float dir):button_(button),state_(state),x_(x),y_(y),dir_(dir){}
 
 	int Button()const{return button_;}
 	int State()const{return state_;}
-	int X()const{return x_;}
-	int Y()const{return y_;}
-	int Dir()const{return dir_;}
+	float X()const{return x_;}
+	float Y()const{return y_;}
+	float Dir()const{return dir_;}
 
 	bool IsClicked()const{return button_>=0 && state_>=0;}
 	bool IsWheel()const{return state_<0;}
